chore: Move where we store our widget implementations
