+++ mobile/src/modules/widget/impl/index.ts
@@ -1,9 +1,9 @@
 // Copyright (C) 2024 - present, MissingCore
 // SPDX-License-Identifier: AGPL-3.0-only
 
-import { ArtworkPlayerWidget } from "../ArtworkPlayerWidget";
-import { NowPlayingWidget } from "../NowPlayingWidget";
-import { ResizableNowPlayingWidget } from "../ResizableNowPlayingWidget";
+import { ArtworkPlayerWidget } from "./ArtworkPlayerWidget";
+import { NowPlayingWidget } from "./NowPlayingWidget";
+import { ResizableNowPlayingWidget } from "./ResizableNowPlayingWidget";
 
 export const nameToWidget = {
   ArtworkPlayer: ArtworkPlayerWidget,
